Removed redundant print statements, fixed waste metric
* Push debugging logic


* Push debugging logic


* Push debugging logic


* Push debugging logic


* added label for algo


* added label for algo


* Debugging for errors


* added a new algorithm for least target change


* Improved the syntax


* Error handling


* Removed the println error statement


* Added verbose comments


* Insufficient test passes


* Changed the order to waste first then lowest change


* Changed the order to waste first then lowest change


* Cleanups for release


* Unstable knapsack, issue in fee calculation which will give fail assertions for high fee rates


* Stress testing LC for fees


* Stress testing BNB for fees


* Stress testing BNB for fees


* Stress testing FIFO for fees


* Stress testing knapsack for fees


* Stress testing older knapsack for fees


* Stress testing LL for fees


* Updates to LC


* Updates to LC


* Updates to Knapsack


* Cleanups


* Lint

* Cleanups


* Cleanups


* Fixed and unified the fee mechanism, simplified the algorithms, changes in test


* Error handling for zero target


* Updates to temp. correct impl of waste metric


* Removed the print statement


* Lints and fixes


---------

+++ src/algorithms/bnb.rs
@@ -10,7 +10,7 @@ struct BnbContext {
     match_range: u64,
     options: CoinSelectionOpt,
     tries: u32,
-    best_solution: Option<(Vec<usize>, u64)>,
+    best_solution: Option<(Vec<usize>, f32)>,
     // Used as a solution to Clippy's `Too Many Arguments` Warn.
     // https://rust-lang.github.io/rust-clippy/master/#too_many_arguments
 }
